Extract Aspire.Hosting.Dapr.Tests
