+++ src/Aspire.Hosting/Aspire.Hosting.csproj
@@ -68,6 +68,7 @@
   <ItemGroup>
     <InternalsVisibleTo Include="Aspire.Hosting.Tests" />
     <InternalsVisibleTo Include="Aspire.Hosting.Containers.Tests" />
+    <InternalsVisibleTo Include="Aspire.Hosting.Dapr.Tests" />
   </ItemGroup>
 
 </Project>
