+++ tests/Aspire.Hosting.Dapr.Tests/DaprSchemaTests.cs
@@ -1,0 +1,30 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Hosting.Tests.Schema;
+using Xunit;
+
+namespace Aspire.Hosting.Dapr.Tests;
+
+public class AzureSchemaTests
+{
+    [Fact]
+    public void ValidateDaprWithComponentsSamples()
+    {
+        new SchemaTests().ValidateApplicationSamples("DaprWithComponents", (IDistributedApplicationBuilder builder) =>
+            {
+                var dapr = builder.AddDapr(dopts =>
+                {
+                    // Just to avoid dynamic discovery which will throw.
+                    dopts.DaprPath = "notrealpath";
+                });
+                var state = dapr.AddDaprStateStore("daprstate");
+                var pubsub = dapr.AddDaprPubSub("daprpubsub");
+
+                builder.AddProject<Projects.ServiceA>("project")
+                       .WithDaprSidecar()
+                       .WithReference(state)
+                       .WithReference(pubsub);
+        });
+    }
+}
