+++ tests/Aspire.Hosting.Tests/Schema/SchemaTests.cs
@@ -151,23 +151,6 @@ public class SchemaTests
                         builder.AddAWSCloudFormationTemplate("TemplateStack", "nonexistenttemplate")
                                .WithReference(awsSdkConfig);
                     }
-                },
-
-                { "DaprWithComponents", (IDistributedApplicationBuilder builder) =>
-                    {
-                        var dapr = builder.AddDapr(dopts =>
-                        {
-                            // Just to avoid dynamic discovery which will throw.
-                            dopts.DaprPath = "notrealpath";
-                        });
-                        var state = dapr.AddDaprStateStore("daprstate");
-                        var pubsub = dapr.AddDaprPubSub("daprpubsub");
-
-                        builder.AddProject<Projects.ServiceA>("project")
-                               .WithDaprSidecar()
-                               .WithReference(state)
-                               .WithReference(pubsub);
-                    }
                 }
             };
